LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audiobuffersourcenode-interface/buffer-resampling-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "interpolate" 
PASS Audit report 
PASS > [interpolate] Interpolation of AudioBuffers to context sample rate 
FAIL X Interpolated sine wave does not equal [0,0.057564035058021545,0.11493717133998871,0.17192906141281128,0.22835084795951843,0.2840152382850647,0.3387378752231598,0.39233696460723877,0.44463497400283813,0.4954586327075958,0.5446387529373169,0.5920130610466003,0.6374237537384033,0.6807206869125366,0.7217600345611572,0.7604058980941772...] with an element-wise tolerance of {"absoluteThreshold":0.090348,"relativeThreshold":0}.
	Index	Actual			Expected		AbsError		RelError		Test threshold
	[176]	-5.6208336353302002e-1	-6.5342020988464355e-1	9.1336846351623535e-2	1.3978270792657052e-1	9.0347999999999998e-2
	[177]	-5.6208336353302002e-1	-6.9591248035430908e-1	1.3382911682128906e-1	1.9230739582821221e-1	9.0347999999999998e-2
	[178]	-5.6208336353302002e-1	-7.3609656095504761e-1	1.7401319742202759e-1	2.3639995980453213e-1	9.0347999999999998e-2
	[179]	-5.6208336353302002e-1	-7.7383983135223389e-1	2.1175646781921387e-1	2.7364379454231952e-1	9.0347999999999998e-2
	Max AbsError of 2.1175646781921387e-1 at index of 179.
	Max RelError of 2.7364379454231952e-1 at index of 179.
 assert_true: expected true got false
FAIL X SNR (32.76 dB) is not greater than or equal to 37.17. Got 32.762424359894766. assert_true: expected true got false
FAIL < [interpolate] 2 out of 2 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 1 tasks were failed. assert_true: expected true got false
